Move query_model to query_model_old
Using:
```
find . -name '*.py' | xargs sed -ri 's/\bquery_model\b/query_model_old/g'
isort .
git mv databuilder/query_model.py databuilder/query_model_old.py
git mv tests/test_query_model.py tests/test_query_model_old.py
```
